// src/app/models/yard-management/truck-arrival.model.ts
export class TruckArrivalModel {
  id: number = 0;
  branch_id: number = 0;
  ta_number: number = 0;
  ta_date: any = new Date();
  manual_doc_number: number = 0;
  customer_id: number = 0;
  customer: string = '';
  address: string = '';
  tr_id: any = 0;
  tr_number: string = '';
  container_number: string = '';
  trucking_company: string = '';
  container_size: string = '';
  container_type: string = '';
  temperature: string = '';
  arrival_time: any = new Date();
  arrival_delivery_no: string = '';
  arrival_delivery_type: string = '';
  arrival_seal_no: string = '';
  arrival_driver: string = '';
  arrival_helper: string = '';
  departure_time: any = '0000-00-00 00:00:00';
  departure_delivery_no: number = 0;
  departure_delivery_type: string = '';
  departure_seal_no: number = 0;
  departure_driver: string = '';
  departure_helper: string = '';
  total_number_of_hours: number = 0;
  warehouse_loading_dock_id: any = 0;
  loading_dock: string = '';
  yard_plugin_station_id: any = 0;
  plugin_station: string = '';
  remarks: string = '';
  status: string = '';
  prepared_by_user_id: number = 0;
  checked_by_user_id: number = 0;
  approved_by_user_id: number = 0;
  released_by_user_id: number = 0;
  is_locked: boolean = false;
}